+++ src/compiler/compiler_internal.h
@@ -2290,7 +2290,7 @@ bool sema_analyse_statement(SemaContext *context, Ast *statement);
 bool sema_expr_analyse_assign_right_side(SemaContext *context, Expr *expr, Type *left_type, Expr *right,
                                          bool is_unwrapped_var, bool is_declaration);
 bool sema_expr_analyse_initializer_list(SemaContext *context, Type *to, Expr *expr);
-Expr **sema_expand_vasplat_exprs(SemaContext *c, Expr **exprs);
+Expr **sema_expand_vasplat_exprs(SemaContext *context, Expr **exprs);
 
 bool sema_expr_analyse_general_call(SemaContext *context, Expr *expr, Decl *decl, Expr *struct_var, bool optional,
                                     bool *no_match_ref);